Rename a database from db toMovieRecSystem also add MyBatis configuration with MyBatisUtil and config.properties
