+++ src/main/java/org/solvd/recommendation/Main.java
@@ -1,7 +1,11 @@
-package org.example;
+package org.solvd.recommendation;
+
+import org.apache.ibatis.session.SqlSessionFactoryBuilder;
 
 public class Main {
     public static void main(String[] args) {
         System.out.println("Hello world!");
+
+
     }
 }
