+++ src/main/resources/config.properties
@@ -1,0 +1,4 @@
+db.user=root
+db.password=devintern
+db.url=jdbc:mysql://34.205.43.78:3306/MovieRecSystem
+mysql.driver=com.mysql.cj.jdbc.Driver
